+++ CHANGELOG.md
@@ -97,6 +97,8 @@ different versioning scheme, following the Haskell community's
     (drops the `std::shared_ptr`). Also the `unary_call_result` now exposes
     read-only getters rather than fields.
   - The `bond::ext::gRPC::wait_callback::arg_type` has been removed.
+  - The `bond::ext::gRPC::server_builder::BuildAndStart` now returns a plain
+    `bond::ext::gRPC::server` object.
   - The `bond::ext::gRPC::server_builder::RegisterService` now only accepts service
     instances managed by `std::unique_ptr`. This properly models the lifetime requirements.
 * gRPC v1.10.0 is now required to use Bond-over-gRPC.
@@ -131,6 +133,7 @@ different versioning scheme, following the Haskell community's
 * Fixed a bug in `bond::make_box` where `const T&` was not handled correctly.
 * The use of `bond::check_method` has been replaced with less restricting expression
   SFINAE checks on supported compilers. [Issue #896](https://github.com/Microsoft/bond/issues/896)
+* Fixed a bug where `bond::ext::gRPC::io_manager` could cause a thread to join itself.
 
 ### C# ###